move app to app.py
